remove fontconfig module within fontconfig-sys crate
It was confusing to have a fontconfig crate which depends on the
fontconfig-sys crate which in turn has a module called fontconfig.
This simplifies the paths so fontconfig-sys does not have any
submodules; everything is in its lib.rs now. So import paths change
from

use fontconfig_sys::fontconfig::*

to

use fontconfig_sys::*
